# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ThetaFrame API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /daily-frames:
    get:
      operationId: listDailyFrames
      tags: [daily-frames]
      summary: List all daily frames for the current user
      responses:
        "200":
          description: List of daily frames
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyFrame"
    post:
      operationId: createDailyFrame
      tags: [daily-frames]
      summary: Create or update a daily frame (date in body)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDailyFrameBody"
      responses:
        "200":
          description: Created or updated daily frame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyFrame"
  /daily-frames/{date}:
    get:
      operationId: getDailyFrame
      tags: [daily-frames]
      summary: Get daily frame for a date
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            description: Date in YYYY-MM-DD format
      responses:
        "200":
          description: Daily frame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyFrame"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertDailyFrame
      tags: [daily-frames]
      summary: Create or update a daily frame
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertDailyFrameBody"
      responses:
        "200":
          description: Upserted daily frame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyFrame"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    TierTask:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        completed:
          type: boolean
        emoji:
          type: ["string", "null"]
      required:
        - id
        - text
        - completed
    TimeBlock:
      type: object
      properties:
        id:
          type: string
        startTime:
          type: string
          description: HH:MM format
        action:
          type: string
      required:
        - id
        - startTime
        - action
    DailyFrame:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        date:
          type: string
          description: YYYY-MM-DD
        colourState:
          type: string
          enum: [green, yellow, red, blue, purple]
        tierA:
          type: array
          items:
            $ref: "#/components/schemas/TierTask"
        tierB:
          type: array
          items:
            $ref: "#/components/schemas/TierTask"
        timeBlocks:
          type: array
          items:
            $ref: "#/components/schemas/TimeBlock"
        microWin:
          type: ["string", "null"]
        skipProtocolUsed:
          type: boolean
        skipProtocolChoice:
          type: ["string", "null"]
          enum: ["micro-win", "intentional-recovery", null]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - date
        - colourState
        - tierA
        - tierB
        - timeBlocks
        - skipProtocolUsed
    CreateDailyFrameBody:
      type: object
      properties:
        date:
          type: string
          description: Date in YYYY-MM-DD format
        colourState:
          type: string
          enum: [green, yellow, red, blue, purple]
        tierA:
          type: array
          items:
            $ref: "#/components/schemas/TierTask"
        tierB:
          type: array
          items:
            $ref: "#/components/schemas/TierTask"
        timeBlocks:
          type: array
          items:
            $ref: "#/components/schemas/TimeBlock"
        microWin:
          type: ["string", "null"]
        skipProtocolUsed:
          type: boolean
        skipProtocolChoice:
          type: ["string", "null"]
          enum: ["micro-win", "intentional-recovery", null]
      required:
        - date
        - colourState
        - tierA
        - tierB
        - timeBlocks
        - skipProtocolUsed
    UpsertDailyFrameBody:
      type: object
      properties:
        colourState:
          type: string
          enum: [green, yellow, red, blue, purple]
        tierA:
          type: array
          items:
            $ref: "#/components/schemas/TierTask"
        tierB:
          type: array
          items:
            $ref: "#/components/schemas/TierTask"
        timeBlocks:
          type: array
          items:
            $ref: "#/components/schemas/TimeBlock"
        microWin:
          type: ["string", "null"]
        skipProtocolUsed:
          type: boolean
        skipProtocolChoice:
          type: ["string", "null"]
          enum: ["micro-win", "intentional-recovery", null]
      required:
        - colourState
        - tierA
        - tierB
        - timeBlocks
        - skipProtocolUsed
...